+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/models/IvyModule.kt
@@ -76,7 +76,7 @@ data class IvyModule(
  * As we remove the drive letter, we later have to guess which drive the artifact belongs to by iterating over A:/, B:/, C:/, ...
  * but that's not a huge problem.
  *
- * @see IntelliJPlatformRepositoriesExtension.localPlatformArtifacts
+ * @see IntelliJPlatformRepositoriesExtension.jetBrainsCdn
  */
 internal fun Path.toPublication() = IvyModule.Publication(
     name = invariantSeparatorsPathString.replaceFirst(Regex("^[a-zA-Z]:/"), "/"),